+++ src/components/containers/HeroContainer.jsx
@@ -3,7 +3,13 @@ import background from '../../images/heroImages/hero-background.png';
 import Header from './Header';
 import NavBar from './NavBar';
 
-const heroBackground = {
+// const homeHeroBackground = {
+//   backgroundImage: `url(${background})`,
+//   backgroundRepeat: 'no-repeat',
+//   backgroundPositionX: 'right',
+// };
+
+const menuHeroBackground = {
   backgroundImage: `url(${background})`,
   backgroundRepeat: 'no-repeat',
   backgroundPositionX: 'right',
@@ -11,7 +17,7 @@ const heroBackground = {
 
 export default function HeroContainer() {
   return (
-    <div className="pt-[57px] pl-[101px] pr-[94px] pb-[185px] w-[100%] min-h[0px]" style={heroBackground}>
+    <div className="pt-[57px] pl-[101px] pr-[94px] pb-[185px] w-[100%] min-h[0px]" style={menuHeroBackground}>
       <NavBar />
       <Header />
     </div>
