+++ src/components/pages/Menu.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Menu() {
+  return (
+    <h1>Menu page</h1>
+  );
+}
